Refactory, adicionando lógica para edição.
