+++ src/app/app-routing.module.ts
@@ -2,11 +2,13 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { SobreComponent } from './layout/sobre/sobre.component';
 import { CardComponent } from './layout/card/card.component';
+import { CardDetailComponent } from './layout/card-detail/card-detail.component';
 
 const routes: Routes = [
   
   {path: 'sobre', component: SobreComponent},
   {path: 'home', component: CardComponent},
+  {path: 'filme/:id', component: CardDetailComponent},
   {path: '', redirectTo: 'home', pathMatch: 'full'}
 ];
