+++ src/app/layout/card-detail/card-detail.component.html
@@ -1,0 +1,9 @@
+<div class="container">
+    <h5>Detalhe do Filme</h5>
+    <div *ngIf="filme">
+        <p>{{filme.titulo}}</p>
+        <p>{{filme.ano}}</p>
+    </div>
+</div>
+
+
